+++ crates/hyperswitch_interfaces/src/lib.rs
@@ -3,6 +3,8 @@
 
 pub mod api;
 pub mod configs;
+/// definition of the new connector integration trait
+pub mod connector_integration_v2;
 pub mod encryption_interface;
 pub mod errors;
 pub mod events;
