+++ crates/router/src/consts.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "olap")]
 pub mod user;
 pub mod user_role;
-
+pub use common_utils::consts::{NO_ERROR_CODE, NO_ERROR_MESSAGE};
 // ID generation
 pub(crate) const ID_LENGTH: usize = 20;
 pub(crate) const MAX_ID_LENGTH: usize = 64;
@@ -31,8 +31,6 @@ pub const DEFAULT_SESSION_EXPIRY: i64 = 15 * 60;
 pub const FINGERPRINT_SECRET_LENGTH: usize = 64;
 
 // String literals
-pub(crate) const NO_ERROR_MESSAGE: &str = "No error message";
-pub(crate) const NO_ERROR_CODE: &str = "No error code";
 pub(crate) const UNSUPPORTED_ERROR_MESSAGE: &str = "Unsupported response type";
 pub(crate) const LOW_BALANCE_ERROR_MESSAGE: &str = "Insufficient balance in the payment method";
 pub(crate) const CONNECTOR_UNAUTHORIZED_ERROR: &str = "Authentication Error from the connector";
