+++ crates/router/src/services.rs
@@ -10,15 +10,16 @@ pub mod pm_auth;
 #[cfg(feature = "recon")]
 pub mod recon;
 
-pub mod connector_integration_v2;
 #[cfg(feature = "email")]
 pub mod email;
 
 use std::sync::Arc;
 
-pub use connector_integration_v2::*;
 use error_stack::ResultExt;
 use hyperswitch_domain_models::errors::StorageResult;
+pub use hyperswitch_interfaces::connector_integration_v2::{
+    BoxedConnectorIntegrationV2, ConnectorIntegrationAnyV2, ConnectorIntegrationV2,
+};
 use masking::{ExposeInterface, StrongSecret};
 #[cfg(feature = "kv_store")]
 use storage_impl::KVRouterStore;
